apps/api/pyproject.toml
[project]
name = "legal-agentic-rag-api"
version = "0.1.0"
description = "Runtime API for Legal Agentic RAG platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "python-multipart>=0.0.20",
  "pypdf>=5.4.0",
  "psycopg[binary]>=3.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0"
]
